+++ spec/app/models/product.rb
@@ -1,6 +1,7 @@
 class Product
   include Mongoid::Document
   field :description, :localize => true
+  field :name, :localize => true, :default => "no translation"
   field :price, :type => Integer
   alias_attribute :cost, :price
 end